updated email verification link to user our domain
