+++ backend/package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "test": "node index.test.js"
   },
-  "author": "Lakshya Gupta, Harsh Jhunjhunwala",
+  "author": "Lakshya Gupta, Aditi Dagar",
   "license": "ISC",
   "dependencies": {
     "aws-sdk": "^2.679.0",
